float->double

+++ Testing/Code/Radiometry/otbMultiChannelGAndRIndexImageFilter.cxx
@@ -58,7 +58,7 @@ int otbMultiChannelGAndRIndexImageFilter(int argc, char * argv[])
 {
   const unsigned int Dimension = 2;
   typedef otb::VectorImage<unsigned char ,Dimension> InputImageType;
-  typedef otb::Image<float,Dimension> OutputImageType;
+  typedef otb::Image<double,Dimension> OutputImageType;
 
   std::string strArgv(argv[1]);
   argc--;
